@@ -7,99 +7,120 @@
     "contributors.add": "all-contributors add",
     "contributors.generate": "all-contributors generate",
     "npm.publish": "nx run-many --target=build --all --parallel=3 && npx concurrently 'npm:npm.publish.*'",
-    "npm.publish.nx-dapr": "npm publish ./dist/packages/nx-dapr/ --access=public",
-    "npm.publish.nx-ddd-architecture": "npm publish ./dist/packages/nx-ddd-architecture/ --access=public",
-    "npm.publish.nx-php": "npm publish ./dist/packages/nx-php/ --access=public",
-    "npm.publish.nx-python": "npm publish ./dist/packages/nx-python/ --access=public",
-    "npm.publish.nx-python-fastapi": "npm publish ./dist/packages/nx-python-fastapi/ --access=public",
+    "npm.publish.nxe-dapr": "npm publish ./dist/packages/nxe-dapr/ --access=public",
+    "npm.publish.nxe-ddd-architecture": "npm publish ./dist/packages/nxe-ddd-architecture/ --access=public",
+    "npm.publish.nxe-php": "npm publish ./dist/packages/nxe-php/ --access=public",
+    "npm.publish.nxe-python": "npm publish ./dist/packages/nxe-python/ --access=public",
+    "npm.publish.nxe-python-fastapi": "npm publish ./dist/packages/nxe-python-fastapi/ --access=public",
     "prepare": "husky install",
     "release.major": "nx run workspace:version --skipProjectChangelog --syncVersions --tagPrefix '' --origin github --releaseAs=major",
     "release.minor": "nx run workspace:version --skipProjectChangelog --syncVersions --tagPrefix '' --origin github --releaseAs=minor",
     "release.patch": "nx run workspace:version --skipProjectChangelog --syncVersions --tagPrefix '' --origin github --releaseAs=patch",
     "sort-package-json": "npx sort-package-json"
   },
-  "resolutions": {
-    "file-saver": "2.0.5",
-    "rxjs": "7.5.7"
-  },
   "dependencies": {
-    "@angular-devkit/core": "^14.2.6",
-    "@angular/core": "~14.2.6",
-    "@docusaurus/core": "2.1.0",
-    "@docusaurus/preset-classic": "2.1.0",
-    "@mdx-js/react": "^2.1.5",
-    "@storybook/builder-webpack5": "~6.5.12",
-    "@storybook/core-server": "~6.5.12",
-    "@storybook/manager-webpack5": "~6.5.12",
+    "@angular-devkit/core": "15.1.4",
+    "@angular/core": "15.1.3",
+    "@docusaurus/core": "2.3.1",
+    "@docusaurus/preset-classic": "2.3.1",
+    "@mdx-js/react": "^2.2.1",
+    "@storybook/builder-webpack5": "6.5.16",
+    "@storybook/core-server": "6.5.16",
+    "@storybook/manager-webpack5": "6.5.16",
+    "@swc/helpers": "~0.4.11",
     "clsx": "^1.2.1",
-    "concurrently": "^7.4.0",
+    "concurrently": "^7.6.0",
     "env-cmd": "^10.1.0",
     "prism-react-renderer": "^1.3.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tslib": "^2.4.0"
+    "tslib": "^2.5.0"
   },
   "devDependencies": {
-    "@angular-architects/module-federation": "^14.3.12",
-    "@angular-devkit/schematics": "14.2.6",
-    "@commitlint/cli": "^17.1.2",
-    "@commitlint/config-angular": "^17.1.0",
-    "@commitlint/config-conventional": "^17.1.0",
-    "@docusaurus/module-type-aliases": "2.1.0",
-    "@jscutlery/semver": "^2.27.1",
-    "@nrwl/angular": "15.0.0",
-    "@nrwl/cli": "15.0.0",
-    "@nrwl/cypress": "15.0.0",
-    "@nrwl/devkit": "15.0.0",
-    "@nrwl/eslint-plugin-nx": "15.0.0",
-    "@nrwl/jest": "15.0.0",
-    "@nrwl/js": "15.0.0",
-    "@nrwl/linter": "15.0.0",
-    "@nrwl/nest": "15.0.0",
-    "@nrwl/node": "15.0.0",
-    "@nrwl/nx-cloud": "14.7.0",
-    "@nrwl/nx-plugin": "15.0.0",
-    "@nrwl/react": "15.0.0",
-    "@nrwl/storybook": "15.0.0",
-    "@nrwl/web": "15.0.0",
-    "@nrwl/workspace": "15.0.0",
+    "@angular-architects/module-federation": "^15.0.3",
+    "@angular-devkit/schematics": "15.1.4",
+    "@commitlint/cli": "^17.4.2",
+    "@commitlint/config-angular": "^17.4.2",
+    "@commitlint/config-conventional": "^17.4.2",
+    "@docusaurus/module-type-aliases": "2.3.1",
+    "@jscutlery/semver": "^2.29.3",
+    "@nrwl/angular": "15.7.0-beta.2",
+    "@nrwl/cli": "15.7.0-beta.2",
+    "@nrwl/cypress": "15.7.0-beta.2",
+    "@nrwl/devkit": "15.7.0-beta.2",
+    "@nrwl/eslint-plugin-nx": "15.7.0-beta.2",
+    "@nrwl/jest": "15.7.0-beta.2",
+    "@nrwl/js": "15.7.0-beta.2",
+    "@nrwl/linter": "15.7.0-beta.2",
+    "@nrwl/nest": "15.7.0-beta.2",
+    "@nrwl/node": "15.7.0-beta.2",
+    "@nrwl/nx-cloud": "15.0.3",
+    "@nrwl/nx-plugin": "15.7.0-beta.2",
+    "@nrwl/react": "15.7.0-beta.2",
+    "@nrwl/storybook": "15.7.0-beta.2",
+    "@nrwl/web": "15.7.0-beta.2",
+    "@nrwl/workspace": "15.7.0-beta.2",
     "@nx-clean/plugin-core": "^1.7.0",
-    "@nx-dotnet/core": "^1.14.0",
+    "@nx-dotnet/core": "1.19.1",
     "@nx-go/nx-go": "^2.7.0",
-    "@nx-plus/docusaurus": "^14.1.0",
-    "@nx-tools/nx-docker": "^3.0.4",
+    "@nx-plus/docusaurus": "15.0.0-rc.0",
+    "@nx-tools/nx-container": "^4.0.3",
     "@nxrs/cargo": "^0.3.3",
-    "@swc-node/register": "^1.4.2",
-    "@swc/core": "^1.2.173",
-    "@twittwer/compodoc": "^1.8.2",
+    "@schematics/angular": "~15.1.0",
+    "@swc-node/register": "^1.5.5",
+    "@swc/cli": "~0.1.55",
+    "@swc/core": "^1.3.32",
+    "@trivago/prettier-plugin-sort-imports": "^4.0.0",
+    "@twittwer/compodoc": "^1.9.0",
     "@types/jest": "28.1.8",
-    "@types/node": "18.11.0",
+    "@types/node": "18.11.19",
     "@types/shelljs": "^0.8.11",
-    "@typescript-eslint/eslint-plugin": "5.40.1",
-    "@typescript-eslint/parser": "5.40.1",
+    "@typescript-eslint/eslint-plugin": "5.50.0",
+    "@typescript-eslint/parser": "5.50.0",
     "all-contributors-cli": "^6.24.0",
     "dotenv": "^16.0.3",
-    "esbuild": "^0.15.11",
-    "eslint": "8.25.0",
-    "eslint-config-prettier": "^8.5.0",
-    "husky": "^8.0.1",
+    "esbuild": "^0.17.5",
+    "eslint": "8.33.0",
+    "eslint-config-prettier": "^8.6.0",
+    "husky": "^8.0.3",
     "jest": "28.1.3",
-    "jest-extended": "^3.1.0",
-    "nx": "15.0.0",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-extended": "^3.2.3",
+    "jsonc-eslint-parser": "^2.1.0",
+    "nx": "15.7.0-beta.2",
     "nx-plugin-esbuild": "^0.5.0",
     "pm2": "^5.2.2",
-    "prettier": "2.7.1",
+    "prettier": "2.8.3",
+    "prettier-eslint": "15.0.1",
+    "prettier-plugin-css-order": "^1.3.0",
+    "prettier-plugin-jsdoc": "0.4.2",
+    "prettier-plugin-organize-attributes": "^0.0.5",
+    "prettier-plugin-packagejson": "^2.4.2",
+    "prettier-plugin-pkg": "^0.17.1",
+    "prettier-plugin-sh": "^0.12.8",
+    "prettier-plugin-sort-json": "^1.0.0",
+    "prettier-plugin-style-order": "^0.2.2",
+    "prettier-plugin-tailwindcss": "^0.2.2",
+    "prettier-plugin-two-style-order": "^1.0.1",
+    "rxjs": "7.8.0",
     "shelljs": "^0.8.5",
     "ts-jest": "28.0.8",
     "ts-node": "10.9.1",
-    "tslib": "^2.4.0",
-    "typescript": "4.8.4",
-    "webpack": "^5.74.0"
+    "tslib": "^2.5.0",
+    "typescript": "next",
+    "webpack": "^5.75.0"
   },
-  "packageManager": "pnpm@7.13.0",
+  "packageManager": "pnpm@7.26.0",
   "engines": {
-    "node": ">=18.0.0",
-    "npm": ">=8.0.0",
-    "pnpm": ">=7.13.0"
+    "node": ">=19.5.0",
+    "npm": ">=9.0.0",
+    "pnpm": ">=7.26.0"
+  },
+  "pnpm": {
+    "overrides": {
+      "file-saver": "2.0.5",
+      "rxjs": "7.8.0",
+      "typescript": "next"
+    }
   }
 }
