@@ -1,4 +1,4 @@
-const { getJestProjects } = require('@nrwl/jest');
+import { getJestProjects } from '@nrwl/jest';
 
 export default {
   projects: getJestProjects(),
