@@ -5,12 +5,12 @@ import {
   runNxCommandAsync,
   uniq,
 } from '@nrwl/nx-plugin/testing';
-describe('nx-dapr e2e', () => {
-  it('should create nx-dapr', async () => {
-    const plugin = uniq('nx-dapr');
-    ensureNxProject('@nx-enterprise/nx-dapr', 'dist/packages/nx-dapr');
+describe('nxe-dapr e2e', () => {
+  it('should create nxe-dapr', async () => {
+    const plugin = uniq('nxe-dapr');
+    ensureNxProject('@nx-enterprise/nxe-dapr', 'dist/packages/nxe-dapr');
     await runNxCommandAsync(
-      `generate @nx-enterprise/nx-dapr:nx-dapr ${plugin}`
+      `generate @nx-enterprise/nxe-dapr:nxe-dapr ${plugin}`
     );
 
     const result = await runNxCommandAsync(`build ${plugin}`);
@@ -19,10 +19,10 @@ describe('nx-dapr e2e', () => {
 
   describe('--directory', () => {
     it('should create src in the specified directory', async () => {
-      const plugin = uniq('nx-dapr');
-      ensureNxProject('@nx-enterprise/nx-dapr', 'dist/packages/nx-dapr');
+      const plugin = uniq('nxe-dapr');
+      ensureNxProject('@nx-enterprise/nxe-dapr', 'dist/packages/nxe-dapr');
       await runNxCommandAsync(
-        `generate @nx-enterprise/nx-dapr:nx-dapr ${plugin} --directory subdir`
+        `generate @nx-enterprise/nxe-dapr:nxe-dapr ${plugin} --directory subdir`
       );
       expect(() =>
         checkFilesExist(`libs/subdir/${plugin}/src/index.ts`)
@@ -32,10 +32,10 @@ describe('nx-dapr e2e', () => {
 
   describe('--tags', () => {
     it('should add tags to the project', async () => {
-      const plugin = uniq('nx-dapr');
-      ensureNxProject('@nx-enterprise/nx-dapr', 'dist/packages/nx-dapr');
+      const plugin = uniq('nxe-dapr');
+      ensureNxProject('@nx-enterprise/nxe-dapr', 'dist/packages/nxe-dapr');
       await runNxCommandAsync(
-        `generate @nx-enterprise/nx-dapr:nx-dapr ${plugin} --tags e2etag,e2ePackage`
+        `generate @nx-enterprise/nxe-dapr:nxe-dapr ${plugin} --tags e2etag,e2ePackage`
       );
       const project = readJson(`libs/${plugin}/project.json`);
       expect(project.tags).toEqual(['e2etag', 'e2ePackage']);
