@@ -5,15 +5,15 @@ import {
   runNxCommandAsync,
   uniq,
 } from '@nrwl/nx-plugin/testing';
-describe('nx-ddd-architecture e2e', () => {
-  it('should create nx-ddd-architecture', async () => {
-    const plugin = uniq('nx-ddd-architecture');
+describe('nxe-ddd-architecture e2e', () => {
+  it('should create nxe-ddd-architecture', async () => {
+    const plugin = uniq('nxe-ddd-architecture');
     ensureNxProject(
-      '@nx-enterprise/nx-ddd-architecture',
-      'dist/packages/nx-ddd-architecture'
+      '@nx-enterprise/nxe-ddd-architecture',
+      'dist/packages/nxe-ddd-architecture'
     );
     await runNxCommandAsync(
-      `generate @nx-enterprise/nx-ddd-architecture:nx-ddd-architecture ${plugin}`
+      `generate @nx-enterprise/nxe-ddd-architecture:nxe-ddd-architecture ${plugin}`
     );
 
     const result = await runNxCommandAsync(`build ${plugin}`);
@@ -22,13 +22,13 @@ describe('nx-ddd-architecture e2e', () => {
 
   describe('--directory', () => {
     it('should create src in the specified directory', async () => {
-      const plugin = uniq('nx-ddd-architecture');
+      const plugin = uniq('nxe-ddd-architecture');
       ensureNxProject(
-        '@nx-enterprise/nx-ddd-architecture',
-        'dist/packages/nx-ddd-architecture'
+        '@nx-enterprise/nxe-ddd-architecture',
+        'dist/packages/nxe-ddd-architecture'
       );
       await runNxCommandAsync(
-        `generate @nx-enterprise/nx-ddd-architecture:nx-ddd-architecture ${plugin} --directory subdir`
+        `generate @nx-enterprise/nxe-ddd-architecture:nxe-ddd-architecture ${plugin} --directory subdir`
       );
       expect(() =>
         checkFilesExist(`libs/subdir/${plugin}/src/index.ts`)
@@ -38,13 +38,13 @@ describe('nx-ddd-architecture e2e', () => {
 
   describe('--tags', () => {
     it('should add tags to the project', async () => {
-      const plugin = uniq('nx-ddd-architecture');
+      const plugin = uniq('nxe-ddd-architecture');
       ensureNxProject(
-        '@nx-enterprise/nx-ddd-architecture',
-        'dist/packages/nx-ddd-architecture'
+        '@nx-enterprise/nxe-ddd-architecture',
+        'dist/packages/nxe-ddd-architecture'
       );
       await runNxCommandAsync(
-        `generate @nx-enterprise/nx-ddd-architecture:nx-ddd-architecture ${plugin} --tags e2etag,e2ePackage`
+        `generate @nx-enterprise/nxe-ddd-architecture:nxe-ddd-architecture ${plugin} --tags e2etag,e2ePackage`
       );
       const project = readJson(`libs/${plugin}/project.json`);
       expect(project.tags).toEqual(['e2etag', 'e2ePackage']);
